ci: use created instead of released as release event type
This change updates the release workflow to use the created event type instead of the released event type. This means that the workflow will now be triggered when a new release is created, rather than when a release is published.

This change is necessary because the released event type is deprecated and will be removed in a future version of GitHub Actions.